+++ app/src/main/java/com/tencent/sgz/bean/Post.java
@@ -45,7 +45,7 @@ public class Post extends Entity{
 	private String face;
 	private String body;
 	private String author;
-	private int authorId;
+	private long authorId;
 	private int answerCount;
 	private int viewCount;
 	private String pubDate;
@@ -109,10 +109,10 @@ public class Post extends Entity{
 	public void setAuthor(String author) {
 		this.author = author;
 	}
-	public int getAuthorId() {
+	public long getAuthorId() {
 		return authorId;
 	}
-	public void setAuthorId(int authorId) {
+	public void setAuthorId(long authorId) {
 		this.authorId = authorId;
 	}
 	public String getFace() {
