+++ app/src/main/java/com/tencent/sgz/ui/NewsDetail.java
@@ -112,7 +112,7 @@ public class NewsDetail extends BaseActivity {
 
 	private int _catalog;
 	private int _id;
-	private int _uid;
+	private long _uid;
 	private String _content;
 	private int _isPostToMyZone;
 
@@ -480,7 +480,7 @@ public class NewsDetail extends BaseActivity {
 				UIHelper.showLoginDialog(NewsDetail.this);
 				return;
 			}
-			final int uid = ac.getLoginUid();
+			final long uid = ac.getLoginUid();
 
 			final Handler handler = new Handler() {
 				public void handleMessage(Message msg) {
@@ -633,7 +633,7 @@ public class NewsDetail extends BaseActivity {
 
 						final AppContext ac = (AppContext) getApplication();
 						// 操作--回复 & 删除
-						int uid = ac.getLoginUid();
+						long uid = ac.getLoginUid();
 						// 判断该评论是否是当前登录用户发表的：true--有删除操作 false--没有删除操作
 						if (uid == com.getAuthorId()) {
 							final Handler handler = new Handler() {
