+++ app/src/main/java/com/tencent/sgz/ui/TweetDetail.java
@@ -103,7 +103,7 @@ public class TweetDetail extends BaseActivity {
 
 	private int _catalog;
 	private int _id;
-	private int _uid;
+	private long _uid;
 	private String _content;
 	private int _isPostToMyZone;
 
@@ -336,7 +336,7 @@ public class TweetDetail extends BaseActivity {
 
 						final AppContext ac = (AppContext) getApplication();
 						// 操作--回复 & 删除
-						int uid = ac.getLoginUid();
+						long uid = ac.getLoginUid();
 						// 判断该评论是否是当前登录用户发表的：true--有删除操作 false--没有删除操作
 						if (uid == com.getAuthorId()) {
 							final Handler handler = new Handler() {
